# docs/swagger.yaml
basePath: /
definitions:
...
  doc.Cluster:
    properties:
      active:
        type: boolean
      cloud_storage: {}
      cluster_request:
        $ref: '#/definitions/doc.ClusterRequest'
      cluster_request_id:
        type: integer
      color:
        type: string
      configPath:
        type: string
      context:
        type: string
      createdat:
        type: string
      dns:
        $ref: '#/definitions/doc.DNS'
      dns_id:
        type: integer
      id:
        type: integer
      image_registry:
        $ref: '#/definitions/doc.ImageRegistry'
      image_registry_id:
        type: integer
      labels:
        type: string
      name:
        type: string
      nodes:
        type: integer
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      provider:
        type: string
      provision_percentage:
        type: number
      pv_capacity:
        type: integer
      region:
        type: string
      storage_class:
        type: string
      total_memory:
        type: integer
      weight:
        type: integer
      zone:
        type: string
    type: object
  doc.ClusterRequest:
    properties:
      active:
        type: boolean
      cluster:
        $ref: '#/definitions/doc.Cluster'
      cluster_id:
        type: integer
      cluster_name:
        type: string
      cluster_version:
        type: string
      createdat:
        type: string
      id:
        type: integer
      network_cidr:
        type: string
      network_policy:
        type: boolean
      nfs_detail: {}
      node_group_count:
        type: integer
      node_group_detail: {}
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      provider:
        type: string
      provider_name:
        type: string
      pvc_write_many:
        type: boolean
      region:
        type: string
      regional_cluster:
        type: boolean
      remove_default_node_pool:
        type: boolean
      status:
        type: string
      tls:
        type: string
      type:
        type: string
      vpc_name:
        type: string
      zone: {}
    type: object
  doc.ClusterScopeEnum:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - SHARED_SCOPE
    - ORGANIZATION_SCOPE
...
  doc.DNS:
    properties:
      active:
        type: boolean
      base_domain:
        type: string
      createdat:
        type: string
      id:
        type: integer
      name:
        type: string
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      project_id:
        type: string
      provider:
        type: string
      region:
        type: string
      tls:
        type: string
      zone_id:
        type: string
    type: object
...
  doc.ImageRegistry:
    properties:
      active:
        type: boolean
      createdat:
        type: string
      id:
        type: integer
      name:
        type: string
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      project_name:
        type: string
      provider:
        type: string
      service:
        type: string
      user_name:
        type: string
    type: object
...
  doc.LoadBalancer:
    properties:
      cluster:
        $ref: '#/definitions/doc.Cluster'
      cluster_id:
        type: integer
      createdat:
        type: string
      custom_domain:
        type: string
      id:
        type: integer
      name:
        type: string
      project:
        $ref: '#/definitions/doc.Project'
      project_id:
        type: integer
    type: object
  doc.LoadBalancerStatusResponse:
    properties:
      createdat:
        type: string
      id:
        type: integer
      namespace:
        type: string
      service:
        additionalProperties: true
        type: object
      status:
        type: string
    type: object
...
  doc.Organization:
    properties:
      createdat:
        type: string
      description:
        type: string
      domain:
        type: string
      id:
        type: integer
      image:
        type: string
      members:
        items:
          $ref: '#/definitions/doc.OrganizationMembers'
        type: array
      name:
        type: string
      organization_plan:
        $ref: '#/definitions/doc.OrganizationPlan'
      organization_plan_id:
        type: integer
      plugins:
        items:
          $ref: '#/definitions/doc.Plugin'
        type: array
      user:
        $ref: '#/definitions/doc.User'
      user_id:
        type: integer
    type: object
  doc.OrganizationMembers:
    properties:
      createdat:
        type: string
      id:
        type: integer
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      user:
        $ref: '#/definitions/doc.User'
      user_id:
        type: integer
      user_role:
        $ref: '#/definitions/doc.UserRoleEnum'
    type: object
  doc.OrganizationPlan:
    properties:
      active:
        type: boolean
      attributes:
        type: string
      cluster:
        type: integer
      cores:
        type: integer
      createdat:
        type: string
      id:
        type: integer
      memory:
        type: integer
      name:
        type: string
      no_of_user:
        type: integer
      price:
        type: integer
      weight:
        type: integer
    type: object
...
  doc.Plugin:
    properties:
      active:
        type: boolean
      add_ons:
        items:
          $ref: '#/definitions/doc.Plugin'
        type: array
      attributes:
        type: string
      createdat:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      is_add_on:
        type: boolean
      min_cpu:
        type: integer
      min_memory:
        type: integer
      name:
        type: string
      plugin_category:
        items:
          $ref: '#/definitions/doc.PluginCategory'
        type: array
      service_detail: {}
      source_url:
        type: string
      support_ci:
        type: boolean
    type: object
  doc.PluginCategory:
    properties:
      createdat:
        type: string
      description:
        type: string
      id:
        type: integer
      is_add_on:
        type: boolean
      name:
        type: string
      plugins:
        items:
          $ref: '#/definitions/doc.Plugin'
        type: array
    type: object
...
  doc.Project:
    properties:
      active:
        type: boolean
      cluster_scope:
        $ref: '#/definitions/doc.ClusterScopeEnum'
      createdat:
        type: string
      dedicated_lb:
        type: boolean
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      logging: {}
      monitoring: {}
      name:
        type: string
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      project_code:
        type: string
      region:
        type: string
      subscription:
        $ref: '#/definitions/doc.Subscription'
      subscription_id:
        type: integer
      subscription_updated:
        type: string
      tags: {}
      user:
        $ref: '#/definitions/doc.User'
      user_id:
        type: integer
      variables: {}
    type: object
...
  doc.Subscription:
    properties:
      active:
        type: boolean
      apps:
        type: integer
      attributes:
        type: string
      backups:
        type: integer
      ci_build:
        type: integer
      cores:
        type: integer
      createdat:
        type: string
      cron_job:
        type: integer
      data_transfer:
        type: integer
      disk_space:
        type: integer
      id:
        type: integer
      load_balancer:
        type: integer
      memory:
        type: integer
      name:
        type: string
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      price:
        type: integer
      price_list: {}
      resource_list: {}
      validity:
        type: integer
      weight:
        type: integer
    type: object
  doc.SuccessResponse:
    properties:
      createdat:
        type: string
      id:
        type: integer
      message:
        type: string
    type: object
...
  doc.User:
    properties:
      active:
        type: boolean
      address_updated:
        type: boolean
      company:
        type: string
      createdat:
        type: string
      designation:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: integer
      image:
        type: string
      is_admin:
        type: boolean
      last_name:
        type: string
      quotas:
        $ref: '#/definitions/models.Quotas'
      reference:
        type: string
      used_demo:
        type: boolean
    type: object
...
  doc.UserRoleEnum:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Admin
    - Members
...
  models.Quotas:
    properties:
      user_organization:
        type: integer
      user_project:
        type: integer
    type: object
...
host: api.test.01cloud.dev
info:
  contact:
    email: info@01cloud.com
    name: API Support
  description: 01cloud-api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 01cloud-api
  version: 1.0.1
paths:
...
  /loadBalancer/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a LoadBalancer with LoadBalancer id
      parameters:
      - description: LoadBalancer id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.LoadBalancer'
      security:
      - ApiKeyAuth: []
      summary: Delete LoadBalancer by id
      tags:
      - LoadBalancer
  /loadbalancer:
    post:
      consumes:
      - application/json
      description: Create a new LoadBalancer with the input paylod
      parameters:
      - description: Environment id
        in: path
        name: id
        required: true
        type: integer
      - description: Create LoadBalancer
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.LoadBalancer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/doc.LoadBalancer'
      security:
      - ApiKeyAuth: []
      summary: Create a new LoadBalancer
      tags:
      - LoadBalancer
  /loadbalancer/{id}:
    get:
      consumes:
      - application/json
      description: Get LoadBalancer by id
      parameters:
      - description: LoadBalancer id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.LoadBalancer'
      security:
      - ApiKeyAuth: []
      summary: Get LoadBalancer
      tags:
      - LoadBalancer
  /loadbalancer/{id}/fetch-status:
    get:
      consumes:
      - application/json
      description: Trigger Get LoadBalancer status service by id
      parameters:
      - description: LoadBalancer id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Trigger get LoadBalancer service
      tags:
      - LoadBalancer
  /loadbalancer/{id}/status:
    get:
      consumes:
      - application/json
      description: Get LoadBalancer status by id
      parameters:
      - description: LoadBalancer id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.LoadBalancerStatusResponse'
      security:
      - ApiKeyAuth: []
      summary: Get LoadBalancer status
      tags:
      - LoadBalancer
  /loadbalancers:
    get:
      consumes:
      - application/json
      description: Get array of LoadBalancers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/doc.LoadBalancer'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all LoadBalancers
      tags:
      - LoadBalancer
...
  /project/{id}/loadbalancers:
    get:
      consumes:
      - application/json
      description: Get array of LoadBalancers with the project id
      parameters:
      - description: Project id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/doc.LoadBalancer'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all LoadBalancers
      tags:
      - LoadBalancer
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"